rename: service dto 하위 command 패키지 생성 및 이동
