+++ src/main/java/ddingdong/ddingdongBE/domain/formapplicaion/controller/dto/request/CreateFormApplicationRequest.java
@@ -1,8 +1,8 @@
 package ddingdong.ddingdongBE.domain.formapplicaion.controller.dto.request;
 
 import ddingdong.ddingdongBE.domain.formapplicaion.entity.FormApplicationStatus;
-import ddingdong.ddingdongBE.domain.formapplicaion.service.dto.CreateFormApplicationCommand;
-import ddingdong.ddingdongBE.domain.formapplicaion.service.dto.CreateFormApplicationCommand.CreateFormAnswerCommand;
+import ddingdong.ddingdongBE.domain.formapplicaion.service.dto.command.CreateFormApplicationCommand;
+import ddingdong.ddingdongBE.domain.formapplicaion.service.dto.command.CreateFormApplicationCommand.CreateFormAnswerCommand;
 import io.swagger.v3.oas.annotations.media.ArraySchema;
 import io.swagger.v3.oas.annotations.media.Schema;
 import jakarta.validation.constraints.NotNull;
