+++ src/main/java/ddingdong/ddingdongBE/domain/formapplicaion/service/FacadeUserFormService.java
@@ -1,6 +1,6 @@
 package ddingdong.ddingdongBE.domain.formapplicaion.service;
 
-import ddingdong.ddingdongBE.domain.formapplicaion.service.dto.CreateFormApplicationCommand;
+import ddingdong.ddingdongBE.domain.formapplicaion.service.dto.command.CreateFormApplicationCommand;
 
 public interface FacadeUserFormService {
