+++ src/main/java/ddingdong/ddingdongBE/domain/formapplicaion/service/FacadeUserFormServiceImpl.java
@@ -5,8 +5,8 @@ import ddingdong.ddingdongBE.domain.form.service.FormFieldService;
 import ddingdong.ddingdongBE.domain.form.service.FormService;
 import ddingdong.ddingdongBE.domain.formapplicaion.entity.FormAnswer;
 import ddingdong.ddingdongBE.domain.formapplicaion.entity.FormApplication;
-import ddingdong.ddingdongBE.domain.formapplicaion.service.dto.CreateFormApplicationCommand;
-import ddingdong.ddingdongBE.domain.formapplicaion.service.dto.CreateFormApplicationCommand.CreateFormAnswerCommand;
+import ddingdong.ddingdongBE.domain.formapplicaion.service.dto.command.CreateFormApplicationCommand;
+import ddingdong.ddingdongBE.domain.formapplicaion.service.dto.command.CreateFormApplicationCommand.CreateFormAnswerCommand;
 import java.util.List;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
