+++ src/test/java/ddingdong/ddingdongBE/domain/form/service/FacadeUserFormServiceImplTest.java
@@ -17,8 +17,8 @@ import ddingdong.ddingdongBE.domain.formapplicaion.entity.FormApplication;
 import ddingdong.ddingdongBE.domain.formapplicaion.repository.FormAnswerRepository;
 import ddingdong.ddingdongBE.domain.formapplicaion.repository.FormApplicationRepository;
 import ddingdong.ddingdongBE.domain.formapplicaion.service.FacadeUserFormService;
-import ddingdong.ddingdongBE.domain.formapplicaion.service.dto.CreateFormApplicationCommand;
-import ddingdong.ddingdongBE.domain.formapplicaion.service.dto.CreateFormApplicationCommand.CreateFormAnswerCommand;
+import ddingdong.ddingdongBE.domain.formapplicaion.service.dto.command.CreateFormApplicationCommand;
+import ddingdong.ddingdongBE.domain.formapplicaion.service.dto.command.CreateFormApplicationCommand.CreateFormAnswerCommand;
 import ddingdong.ddingdongBE.domain.user.entity.Role;
 import ddingdong.ddingdongBE.domain.user.entity.User;
 import ddingdong.ddingdongBE.domain.user.repository.UserRepository;
